Refactoring: Don't repeat the lookup and decoding of local TODO_DIRs on each parent directory iteration
Instead, collect the local TODO_DIRs once, and pass them as arguments to the findup command, which is switched from grep to filterArg.
As filterArg succeeds when there's a match and findup stops if the --exec COMMAND fails, we simply have to negate the exit status and thereby achieve a stop of the iteration after the first result.