Renaming: ttl to todo-local.sh, and switch the tt alias to that
Now that ttl acts as a wrapper around todo.sh that either uses the local task list or falls back to the global one, we don't need to mentally differentiate between the two, and just use a single alias for both scopes. (The alias is so deeply embedded in my muscle memory that it felt really odd using the new ttl alias. I think it would take a long time to change that, and that therefore a single unified alias is much better.)