ENH: Generalize todo count to all supported local locations
Not just TODOs in Git repos, by leveraging "todo-local.sh location" as the source for the TODO_DIR.
For efficiency, we store that value in _PS1TodoLocalTodoDir; the lookup is a bit more costly as we cannot simply leverage an existing variable (like _PS1GitRoot) any longer.
The panel function is invoked from a subshell, so we cannot directly memoize the value there, but instead need to define a separate _PS1TodoLocal() event handler for _PS1OnChangeDirectory.