todo-local: Don't treat the global todo.txt as local when we're just in the TODO_DIR
Because the todo.txt is found in the current working directory, we would treat this as a local todo.txt, enable auto-archiving, and redirect DONE_FILE to the user's local data. This means that the global todo list is suddenly treated differently!
I can now easily query the actual todo.sh configuration via my custom config action; with this, it's easy to recognize the global todo.txt and don't treat it as a local one.